Implement basic loadbalancer traffic test
Issues:
Fixes #64

Problem: Need end-to-end traffic test to exercise lbaasv2 driver and agent undergoing active development.  We don't yet have the capabilities to ramp up Tempest, so using an in-house solution to deploy OpenStack.

Analysis: The bulk of the test_solution file is to create a fully-configured proxy for either LBaaSv1 or LBaaSv2, abstrating the underlying lbaas version from the actual traffic test.

Tests: Executes for LBaaSv1 with VXLAN and GRE tunnel types.  Failing for 'none' tunnel when deploying VE into the guest tenant.