Creating loadbalancer with mismatched tenant ids causes error
Issues:
Fixes #78

Problem:
This issue is around the different tenant-ids for the loadbalancer vs.
the network on which it is deployed.  If the two tenant ids are in agreement,
then everything is fine.  If there is a difference between them,
then we get an error when we try to create the tunnel.

So if I have something like the following, it works
TenantA> neutron lbaas-loadbalancer-create —name lb1 tenantA-subnet
Admin> neutron lbaas-loadbalancer-create —name lb1 —tenant-id tenantA-id tenantA-subnet

However, fails:
Admin> neutron lbaas-loadbalancer-create —name lb1 tenantA-subnet

Note that the network for tenantA-subnet is not in the '/Common'
folder

Analysis:
1.  If the network tenant ID matches the loadbalancer objects tenant ID,
then everything is hunky dory.
2.  If the network is a /Common network, (e.g. shared == true or
router:external == true), then the tenant Id of the loadbalancer service
objects need not be the same as the network tenant id.
3.  If the tenant ID¹s of the (non-common) network and loadbalancer
objects don¹t match, then we throw an error that tells the user that
there cannot be a dependency between a loadbalancer tenant object and a
non-common network.

Fixup error handling in driver for loadbalancer creation.

Test:
Manual

fix error in service acquisition