Failing to create loadbalancer on shared network in global routed mode
Issues:
Fixes #227

Problem:
Create a loadbalancer where the loadbalancer tenant id does not
match the network tenant id and global routed mode is enabled.
This fails because the agent configuration does not contain
the 'f5_common_external_networks' item.

Analysis:
In the service builder modify the test to check if a network is a
shared network by testing for the 'f5_common_external_networks'
element in the configuration.

Tests:
Create a loadbalancing service on a shared admin network for a
tenant and pass traffic.