Allow destruction of loadbalancers when there is a tenant id mismatch
Issues:
Fixes #240

Problem:
When a user attempts to create a loadbalancer, but the request fails
b/c the tenant id of the network and the tenant of the loadbalancer
are not the same, a loadbalancer is left in error state. This loadbalancer
cannot be cleared up b/c the same exception is thrown when a delete
request on the loadbalancer is made.

Analysis:
Handle the exception in the loadbalancer delete request so that the
loadbalancer can be deleted.

Tests:
Tempest API v2 tests.