Add tempest tests for CRUD operations for l7policies and rules
Issues:
Fixes #276 #277

Problem:
The existing tempest tests needed to verify the service object is as we
expect after modifications were made to the service builder to add
l7_policies and l7policy_rules to the service object.

Analysis:
Found a couple of bugs during tempest testing and fixed those. Also
found a few issues with tempest setup and teardown that have been fixed
for the l7policy tests, but the l7rule tests are still failing. This is
because we will not adequately mock out agent behavior to get the tests
pass. Instead, we'll wait until the agent behavior is there.

Tests:
All unit tests pass. The l7policy tempest tests pass.