driver_v2 module needs to accommodate detached pools and l7policies and l7rules
Issues:
Fixes #285

Problem:
Current implementation of LBaaSv2 driver assumes the relationships
between lb, listener, and pool objects that lb has one listener and
listener has one pool. A pool must be associated with a lb. This
assumption must be broken in the driver, this refers to work in
driver_v2. It is likely that the work for detached pools is already
done, but we also need to add the CRUD interfaces for l7policies and
l7rules.

Analysis:
Added CRUD for l7policies and l7rules. Also consolidated a great deal of
code to make this module a bit easier to read. The update interface for
all managers, except l7policies and l7rules, remains as it was before,
due to the fact that these rely upon an old entity and the new one with
the change. I left these alone because they will be a part of a further
refactor when @richbrowne completes his naming changes.

Tests:
All new unit tests pass