remove sudo commands, this script must be run as root
Issues:
Fixes #316

Problem: Different kinds of whitespace in the config files was ruining variable extraction.
In order to access the docker socket the container user needs to be root.

Analysis: A better fix that will allow this to run with buildbot might be to change the group
that owns `/var/run/docker.sock` in the container.

Tests:  A successful run of the targets in systest/Makefile will test this commit.