Feature.l7switching requirements file need to have tempest setup
Issues:
Fixes #318

Problem:
The tempest tests need to have a specific version of neutron-lbaas and
tempest to run properly. This is due to the nature of tempest being a
moving target with respect to OpenStack releases. We have found a match
of these two projects at a point in time and we are running tests with
that infrastructure.

Analysis:
Changed the requirements.test.txt file to have a tagged version of
tempest and neutron-lbaas

Tests: