Documentation needed for feature.l7switching release
Issues:
Fixes #332

Problem:
Documentation needed in this repo for feature.l7switching. This will
have explanation of language differences between OpenStack and BIG-IP
and will have a working examples of creating policies and rules.

Analysis:
This doc captures the notes above, and it maps out exactly what the
differences between OpenStack and the BIG-IP are in terms of L7 Content
Switching. It offers working examples as well.

Tests: