Driver make targets duplicated and typo in 11.5.4 undercloud target
Fixes: Issue #386

Problem:
In the nightly builds, the tempest_11.6.0_overcloud make target is duplicated.
It should be changed to tempest_11.6.0_undercloud. Also, a typo exists in the
tempest_11.5.4_overcloud build, where the TEST_OPENSTACK_CLOUD variable is wrong.

Analysis:
Changed the duplicated target to the proper undercloud target and fixed the
typo.

Tests: