Autolog output directory is driving the suite name in trtl results
Fixes: Issue #398

Problem:
When tox is run in nightly, the --autolog-outputdir provided to the tox command causes the suite name to be rooted (and thus named) from the / directory on the worker. We need to provide a .pytest.rootdir file near the tests, so suite name is made relative to the location of the tests, and thus is not filled with redundancies.

Analysis:
Did not commit these files to the repo, but had the .pytest.rootdir files added via the tempest_tests.sh script in systest/scripts. This will prevent the inevitable questions about why these files exist in the public repo.

Tests:
Ran manual tests and manually inspected the names of the suites.