Tempest scenario test bug reporting false positives
Issues:
Fixes #397

Problem:
In the L7 tempest scenario tests we run traffic through the VIP to the
backend servers. We have a retry to ensure we wait for control plane
changes to be applied to the data plane. In that retry, we don't
appropriately check for failure if we're not expecting a failure. We
should raise an exception if we've reached the last retry and we haven't
had a successful connection.

Analysis:
Added a final check in the try/except block in the _run_traffic function
to ensure if we've reached the last iteration and we're still failing,
and we don't expect to fail, then we raise an exception.

Tests:
Tested against a deployment where we expect the scenario tests to fail
(overcloud) and where we expect it to work (undercloud).