Gumballs incapable of parsing DRIVER_TEMPEST_SESSION var in Makefile
Issues:
Fixes #423

Problem:
This fix is currently only applicable to the mitaka branch. We need to
modify the value of the variable above to not contain two underscore
characters, as this isn't acceptable by the gumballs parser.

Analysis:
Fixed the issue by changing the first underscore to a dot character.

Tests: