Fix PROJECT name and _ --> -
Issues:
Fixes #429

Problem: There are two types of data in the GUMBALLS_PROJECT.
The data at the beginning pertains to the code under test, and its
version..  aka the CHANGESOURCE and the BRANCH, the data at the
end pertains the environment under which the code was tested.

Analysis:  We separate these two parts of the name with a '-'.
To delimit the components _within_ a part of the GUMBALLS_PROJECT
we use '_'