In mitaka branch, autolog outputdir needs to be added to tox file
Issues:
Fixes #432

Problem:
The autolog package needs to be installed as part of the tox
dependencies in the tox.ini for the l7 test target. If these tests try
to run, the result will fail if autolog is not installed.

Analysis:
Added the autolog package to the tox.ini file.

Tests: