hardcode target of rm -rf in scripts
Issues:
Fixes #434

Problem: In order to make scripts rerunnable it's desirable to
remove directories, prior to running commands that populate them.
If the targets have variable names it's possible that an error
in variable assignment/creation could cause an incorrect directory
to be removed.

Analysis: This change makes the targets "hard-coded" to prevent
the possibility of a "variable injection" style error.

Tests: The changes were manually tested by running the make targets
in a buildbot sandbox environment.