From_agent_session name in Makefile need to be renamed
Issues:
Fixes #435

Problem:
This change is applicable in liberty and mitaka and is required to make
gumballs report correctly in nightly. The parser which reports results
in gumballs doesn't handle the multiple underscores correctly when
reporting. We need to change one of the underscores to something else,
so those tests will fall in line with others.

Analysis:
Changed the from_agent_session to replace the first underscore with a
dot.

Tests: