Policy object with rules becomes too large for message queue
Issues:
Fixes #449

Problem:
When we retrieve the policy object in the driver_v2.py module with
to_dict, if a rule exists for the policy, the rule will be retrieved
recursively. This has the potential to create a very large policy
object, especially if more than one rule exists.

Analysis:
I fixed this in the to_dict calls in driver_v2, I added a filter on
rules. These will not be recursively evaluated.

Tests:
Ran tests against an 11.6.0 overcloud stack and get passes.