Tempest api tests attempt to delete VIP port, which is managed by loadbalancer
Issues:
Fixes #456

Problem:
The tempest tests often create a loadbalancer in test setup. Once that
lb is created, tempest searches for the VIP port that was also created
by the loadbalancer and puts it in a list to be deleted later.

The test should not need to cleanup that port, as it will be done once
the loadbalancer is deleted. We should have the test poll for the port,
to ensure it was deleted properly. Otherwise, for now, the test will
fail.

Analysis:
Fixed the api base.py module to prevent the port deletion outright, and
properly wait for the loadbalancer delete to subsequently remove the VIP
port.

Tests:
Ran this test against a stack where I modified a sleep on the
loadbalancer delete operation. The sleep was implemented before the VIP
port delete. By varying the time of sleep, we can see what happens when
the loadbalancer deletes the port implicitly and when it doesn't. We
raise if it doesn't.