Driver sets VIP port owner
Issues:
Fixes #461

Problem:
The driver "takes ownership" of the VIP port (which is created by the
load balancer creation operation), and this is unnecessary. This appears
to directly affect the way the port is handled in neutron, especially
upon deletion. We are making this change in the mitaka branch to
troubleshoot nightly failures.

Analysis:
Removed the line to set the owner in driver_v2.py module.

Tests:
Ran overcloud mitaka tests against this change the results look as it
does when not modified. This is as expected, since this issue can crop
up when the test tears down the VIP port at the exact time as the load
balancer deletion operation.