Service builder not using the segment physical_network field
Issues:
Fixes #452

Problem:
Service builder is not completely populating the network dictionary
with all the values from the segments table. The error is that the
phyiscal_network_name is assumed to be defined in the top-level segment,
but this is not always the case.

Analysis:
Add the segment's physical_network field to the network definition
in the service object.

Tests:
Manual testing consisted of:
Create loadbalancer service with all objects: listener, pool, members
Create member with and without existing port.
Create member on opflex, add vlan child segment manually and verify
that the member is connected.
Create loadbalancer on opflex network, add vlan child segment
manual and verify that the virtual get's created.
Create loadbalancer on opflex network, don't add segment and check
that the loadbalacner goes to 'ERROR' state and that it can be deleted.
Create a member on the opflex network, don't add segment and check
that the member goes to 'ERROR' state after timeout period and verify
it can be deleted.