Issues: Fixes #474
Problem:
The issue of the Child PID's (as spawn and used by neutron-server) are
attempting to use the same RPCClient instance to perform communications.

Analysis:
By moving the creation of the RPCClient into the rpc-method call code
on the basis of per-pid, should keep these PID's (parent vs children)
from colliding with one another.

Tests:
The test for this is Nighly's run as a whole