Makefile script in liberty does not teardown barbican containers
Issues:
Fixes #507

Problem:
The makefile script cleanup_tlc_session.sh does not cleanup the barbican
container with the uninstall_barbican tlc command. This is because
previously, we were not using barbican in the liberty branch.

Analysis:
Added the barbican container cleanup to the cleanup_tlc_session.sh
script.

Tests:
Ran the test locally.