Makefile needs to be prepared for newton running in nightly
Issues:
Fixes #579

Problem:
The systest Makefile needs to be able to handle a branch name of
stable/newton. That means we need to sanitized the name to strip out the
forward slash. This is necessary because forward slashes can and will
introduce all sorts of fun issues when that branch name is used in say,
a directory path. Also, jenkins will not allow a job named with a
forward slash. So, we will sanitize that name, then to get the OpenStack
distro, we can strip the 'stable' string out of that sanitized branch
name. This allows us to match the branch name as closely as possible for
internal test reporting and to feed the distro to TLC for testing.

Analysis:
Created a new variable called SANITIZED_BRANCH in the makefile, which is
used to create the project name and to obtain the TEST_OPENSTACK_DISTRO.

Tests:
Testing manually in my jenkins worker.