install_test_infra.sh script needs to clone dev-test repo from unsanitized branch in makefile
Issues:
Fixes #582

Problem:
This only affects newton, but I'm making the change in liberty and
merging it forward so the changes will be consistent across the
branches. The environment variable TEST_OPENSTACK_DISTRO is used as the
branch to clone for the dev-test repo. In the past, this was fine. As of
newton, the branch is now called 'stable/newton', so this will not work,
because that environment variable has been sanitized to be only
'newton'. We need to set the branch properly to the same branch as being
used in driver, because the branches all have the same names across
repos.

Analysis:
Set the branch to be $BRANCH variable exported from Makefile and not the
TEST_OPENSTACK_DISTRO environment variable.

Tests:
Ran tests on local jenkins container.