test_stats test failing in stable/newton due to changes in neutron_lbaas test setup
Issues:
Fixes #609

Problem:
The test_stats scenario test fails in nightly consistently due to
changes made in the parent class of that test. The test expects the
loadbalancer id and the security group id to be attributes of a
loabalancer object and a security group object respectively, but those
are dictionary objects, not lbaasv2 model objects. Only the test needs
to be changed here.

Analysis:
Changed the object attribute accesses to dictionary key accesses for the
security group. For the loadbalancer id, we just use a variable created
previously in the test.

Tests:
Ran test on my local stable/newton jenkins worker and it passes.