Doc references to branch in stable/newton need to be updated
Issues:
Fixes #617

Problem:
For the beta release of stable/newton, the docs should be updated with
references to the newton release, where applicable. There are several
hard-coded releases in our docs files, and this task tracks the
modification of those.

Analysis:
Updated the references of mitaka to newton or stable/newton in the
stable/newton branch. Also bumped the driver version to 10.0.0.b1

Tests:
Tests will run in travis on build.