Service object related to pools and associated listeners needs to be cleaned up
Issues:
Fixes #611

Problem:
A shared pool is one which is associated with multiple listeners. There
is a period of time where the pool is not updated with associated
listeners, and that period of time could overlap with the building of
the service object. If it does, then the pool's list of listeners may
not be accurate. This means the service would be built based on stale
information. The listener, however, does have the correct references to
its associated pool. We will tweak the service object to reflect this
fact. We can have the agent rely upon the listener's view of its pool,
instead of relying upon the pool's view of the listeners. This issue
tracks cleaning up the service object to reflect this change.

Analysis:
Removed the pool's listeners and listener_id from the service object.

Tests:
Create a unit test to ensure we don't add listeners to a pool dict,
nor the listener_id field.