test architecture supports testrunner container
Issues:
Fixes #597

Problem: Our test infrastructure was/is plagued with a large number of sources
of change, many of which are outside of our control.
Among these sources of variation:

* the transitive closure of Python packages our tests depend on (frequently
referenced via requirements.txt files)
* the set of packages built in to the BDO provided test runner
* the root Docker image the test runner was built on
* the community constraints files specified in the neutron-lbaas fork we use to
attempt to match the communities state

There are other known sources of uncontrolled variation, the above is the list
of sources of change that are eliminated by the change to containerized testing
this commit is a part of.

Analysis: In order to stablize our test infrastructure the following
enhancements are implemented:

(1) The test runner build specification is now managed by the F5 OpenStack
Integration Team (us) at:  https://gitlab.pdbld.f5net.com/openstack/labbench

(2) Dependencies are incorporated into the testrunner via Dockerfiles.

(3) Where possible git clone is used to record the source code of dependencies
in the relevant layer of the Dockerfile.  The corresponding application is
installed from source in the same layer.

(4) Management of the testrunner, and other microservices, is implemented with
a simple make-based build system.

These enhancements eliminate the above listed sources of variance.  More work
remains with respect to other sources of variance.

Tests: all the tests in "run_neutron_lbaas" were run to confirm the
functionality of this change.