newton jenkins build should have a long timeout due to more tests being run
Issues:
Fixes #631

Problem:
In the stable/newton branch, the Jenkinsfile specifies a job timeout of
2 hours. We should bump this to 3 hours because nightly runs have been
intermittently failing due to the fact that they hit this timeout. This
makes sense because more tests are being run in newton than in any other
branch.

Analysis:
Updated the timeout to three hours.

Tests:
The timeout test will only run in nightly Jenkins. Since we have no
proxy for this locally, we can let nightly test this for us. The risk is
pretty small.