Retry needed on TLC setup command in setup_tlc_session.sh build script
Issues:
Fixes #608

Problem:
Occasionally, the TLC setup command fails in nightly runs with an error
while scp'ing up scripts to the OpenStack hosts. We should retry the TLC
setup if the rc is anything other than 0. Note that we will have to do
TLC cleanup and then mv the /testlab/sessions/<failed_session_name>
before we retry, because that directory cannot exist when we try to
rerun. Unless, we gave the retry session a new name.

Analysis:
Updated the setup_tlc_session.sh script to retry the setup command if it
fails. The retry limit is three times.

Tests:
Tested on bahamas by running this script directly.