Fixing test_member_subnets test. Selfip check was failing.
Issues:
Fixes #707

Problem:
The api test api.test_member_subnets.MemberSubnetTestJSON fails
consistently with a ValueError. It appears as though the selfip address
and route domain are expected back from a function call, but the route
domain is not used. The function call also does not return a tuple, but
only a single value.

Analysis:
Fixed the expected return value from the function call, so now it only
looks for selfip_address.

Tests:
The test now passes. Ran it on my local mitaka stack.