test_esd.ESDTestJSON.test_create_esd fails with profile not found
Issues:
Fixes #596

Problem:
test_esd fails because ESD is not fully applied before the verification
code runs. Inserting a delay (e.g. time.sleep(3) )before check_esd()
allows test to pass.

Analysis:
Added a retry when checking whether esd has been applied. Since this can
take some time, we may expect one or two iterations of this retry to be
kicked before we succeed in applying esd with the policy given. If we
run out of iterations and we still haven't see the profile we're looking
for, then the test fails.

Tests:
Ran the test multiple times. Did not see a failure ever, but this should
serve as a good starting point for a retry if there's a bug in the test
code.