make the nightly test run start by chowning
Issues:
Fixes #710

Problem:  Nightly TRTL results accumulated extra results because of a
failure to remove directories from the NFS file system.

Analysis:
In order to enable the docker testrunner "jenkins" user to write to the
-v mounted volume it shared with the executor I added "sudo -E" commands
to the systest scripts.  This worked by allowing the write operations
but caused a subsequent bug, the executor could not remove the
sudo-written files.
One fix might be to cause the openstack-test-testrunner container to be
_RUN_ by the "jenkins" user, but the openstack team does not control
that interface.  As a workaround the "jenkins" user _inside_ the
testrunner container, chowns the relevant directory structure at the
start of its test run.

Changes:
* Remove sudo -E, depend on chown -R in Makefile
* remove .PHONY directive, and modify chown directive

Tests:
  This commit was validated by this test run:
http://jenkins.pdbld.f5net.com/job/openstack/job/driver/job/mitaka/job/za_test_11.6.1_undercloud_vxlan/4/