change name from zancas_tmp to container_mailbox
This fixes a temporary file name to be something more reasonable.
The "container_mailbox" has files that the ansiblemicroservice needs to
access FROM the testrunner.  That is the testrunner produces "mail" and
shoves it in the mailbox, then when the microservice runs it receives
that "mail" as input.

I also changed the name of the stub exclude file.

Finally, this commit uses a restricted sudo invocation that avoids
manipulating files on the NFS mount that do not "belong" the openstack
team.