test_pool_sync.py test fixups needed for assuring pool and member are created after manual deletion
Issues:
Fixes #714

Problem:
In the tests at api/test_pool_sync.py, we are checking that manual
deletion (out-of-band of the agent and neutron lbaas) will be rectified
by the agent when it performs a periodic resync or if the service object
where the pool and/or member live is updated after the manual deletion.
These tests need some fixups to get them green.

Analysis:
Updated the expectations of the tests in this module to create the pool
and member in the tests, instead of at setup, that way we have greater
control of the create/delete operations in the tests.

Tests:
Ran the api tempest tests in the driver.