Makefile in stable/newton branch refers to branch as mitaka
Issues:
Fixes #723

Problem:
In what looks like a bad merge conflict resolution, the mitaka branch is
referred to directly in the stable/newton branch. This hard-coding is
necessary for build on PRs, so we can accurately determine the openstack
version when running those tests, irrespective of the PR-creator's
branch name.

Analysis:
Fixed the branch name and will let build on PR run to test this change.

Tests:
Build on PR will run and test that this change works.