test_member_status api tests intermittently fail in nightly tests
Issues:
Fixes #726

Problem:
When the tests in api/test_member_status.py fail, they often fail
together. The reason is that they arbitrarily choose the starting
address in the subnet to use as the member address. This address is
sometimes exactly the same as the VIP address on the loadbalancer.
Neutron does not complain at all, and the driver no longer manages ports
for members. When member is being deleted at the end of the test, the
port is being deleted as well, but we cannot delete the VIP port upon
member deletion.

Analysis:
Using the last address in the subnet. Due to the way neutron-lbaas
allocates IPs for VIps, this is a safer bet than the starting address.

Tests:
Ran the test_member_status tests multiple times to test this change.