L7Policy tests can fail due to timing issue and lb status behavior in mitaka
Issues:
Fixes #626

Problem:
This is a combination of a test issue and an actual issue in
neutron-lbaas (although some might purport it was a design decision in
neutron-lbaas). The intermittent failures of the l7policy tests below
are often caused by the policy not yet existing on the BIG-IP device.
This is because, in the test, the lb provisioning status is used as a
guide for when objects are changing beneath it. This means that if a
listener is created or updated or deleted, the lb's provisioning status
goes into PENDING_UPDATE. Yet when a policy is created (only on the
creation event), the lb's status is not set to PENDING_UPDATE. This is
true for mitaka and newton.

Analysis:
Created new methods to check for existence and non-existence of a
policy. These methods sleep between iterations and only check a few
times.

Tests:
Testing on my local mitaka stack.