ansible_install_lbaasv2.sh should install checkout version of driver code
Issues:
Fixes #733

Problem:
In build on PR, the checked out version of the driver should be used as
the code to install on the OpenStack controller. We just need to change
the variable fed into the ansible_install_lbaasv2.sh script. This
variable will have the driver install location set as the checked out
version of the driver.

Analysis:
Updated the Makefile and the ansible_install_lbaasv2.sh script to get
the current checkout out repo and branch into the ansible container.

Tests:
The build on PR will test this behavior, and this is where we really
want this change to be, so it is a good test.