Change-reverted: Newton process creation event changed from mitaka
Issues:
Fixes #737

Problem:
An issue was fixed previously in stable/newton (then master) which
addressed a change in the process event creation between mitaka and
newton. That change was accidentally reverted in a recent modification.
Now it must be fixed again. The original issue is #530.

Analysis:
Changed the expected process creation event to events.AFTER_INIT, which
is the correct event emitted on process creation in newton.

Tests:
Ran a test on my local newton stack by changing the expected event and
then creating a loadbalancer.