Smoke test failures on PR are no longer causing build failure in Jenkins
Issues:
Fixes #774

Problem:
This may be due to a change in the Makefile, as this was working at one
time. I've made a more obvious change to capture the rc of the test run
and pass it up to the caller in the agent Makefile. I think that change
may be in order here to ensure we produce a non-zero rc if tests fail.

Analysis:
Made the change to capture the rc from the smoke test run. Also, changed
the smoke test that is run on this target invocation.

Tests:
Testing in our jenkins. Will test success and failure modes.