docker pull should be run to get latest ansible container from docker registry
Issues:
Fixes #792

Problem:
A docker pull command should be run before the docker run for the
ansible container used to deploy the heat plugins and the agent/driver.
If we don't pull, we are prone to using a stale container, whichever one
might exist on the docker host. This could produce difficult to debug
errors.

Analysis:
Added a docker pull just before the docker run statement in the
ansible_install_lbaasv2.sh script, so we get the latest container.

Tests:
Will run tests in build on PR.